+++ src/main/java/com/github/funthomas424242/rades/project/domain/RadesProjectAccessor.java
@@ -17,8 +17,13 @@ public interface RadesProjectAccessor {
 
     String getGithubUsername();
 
-    String getGithubRepositoryname() ;
+    String getGithubRepositoryname();
 
+    String getBintrayUsername();
+
+    String getBintrayRepositoryname();
+
+    String getBintrayPackagename();
 
 
 }